+++ kmath-geometry/src/commonMain/kotlin/space/kscience/kmath/geometry/euclidean2d/Circle2D.kt
@@ -1,20 +1,19 @@
 /*
- * Copyright 2018-2022 KMath contributors.
+ * Copyright 2018-2023 KMath contributors.
  * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
  */
 
-package space.kscience.kmath.geometry
+package space.kscience.kmath.geometry.euclidean2d
 
 import kotlinx.serialization.Serializable
-import space.kscience.kmath.geometry.Euclidean2DSpace.distanceTo
-import kotlin.math.*
+import kotlin.math.PI
 
 /**
  * A circle in 2D space
  */
 @Serializable
 public data class Circle2D(
-    @Serializable(Euclidean2DSpace.VectorSerializer::class) public val center: DoubleVector2D,
+    @Serializable(Float64Space2D.VectorSerializer::class) public val center: DoubleVector2D,
     public val radius: Double
 )